+++ src/PlingBot/BotOptions.cs
@@ -4,5 +4,5 @@ public class BotOptions
     public bool TestMode { get; set; }
     public string Game { get; set; } = "Stryktipset";
     public DateOnly? CouponDate { get; set; }
-    
+    public bool IsVersusMode { get; set; }
 }